+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Session` class `unmask_message` method modified
 - `Prompt` class `update_template` method modified
 - `PromptTemplate` class `to_dict` method modified
+- `PromptTemplate` class `render` method modified
 - Test system modified
 - `Response` class `date` bug fixed
 ## [1.2] - 2026-06-15